Add day 21, part 2
This was a tough one - there was a cyclic nature to it, as well as
fitting of a quadratic formula. Had to visit the AoC subreddit for some
hints here.

Additionally, gave up on making the code generalizable to the test cases
provided for part 2 - there was some constraints / parameters that were
valid for the provided input, but not the test input. Instead, used the
test for debugging purposes.